fix(racpach) fix load recpach proccess
