+++ composables/useRecapcha.ts
@@ -1,8 +1,102 @@
-import { useNuxtApp } from "nuxt/app"
+import { useNuxtApp, useRuntimeConfig } from "nuxt/app"
 import type { RecaptchaContext } from '~/plugins/recapcha.client'
+import { ref } from "vue"
 
 export const useRecaptcha = () => {
-    const { $recaptcha } = useNuxtApp()
+    // Get site key from runtime config or environment
+    const siteKey: string = useRuntimeConfig().public.recapchaSiteKey as string
+    const recaptcha = ref<RecaptchaContext> ()
+
+    // Function to dynamically load the reCAPTCHA script
+    function loadRecaptchaScript(): Promise<any> {
+        
+        return new Promise((resolve, reject) => {
+            // Check if grecaptcha is already available
+            if ((window as any).grecaptcha) {
+                return resolve((window as any).grecaptcha)
+            }
+
+            // Check if the script is already present
+            if (document.getElementById('recaptcha-script')) {
+                // Wait for it to load
+                const checkLoaded = () => {
+                    if ((window as any).grecaptcha) {
+                        resolve((window as any).grecaptcha)
+                    } else {
+                        setTimeout(checkLoaded, 100)
+                    }
+                }
+                checkLoaded()
+                return
+            }
+
+            // Create the script tag for reCAPTCHA with the site key
+            const script: HTMLScriptElement = document.createElement('script')
+            script.id = 'recaptcha-script'
+            script.src = `https://www.google.com/recaptcha/api.js?render=${siteKey}`
+            script.async = true
+            script.defer = true
+
+            script.onload = (): void => {
+                // Wait a bit for grecaptcha to be available
+                const checkReady = () => {
+                    if ((window as any).grecaptcha && (window as any).grecaptcha.ready) {
+                        (window as any).grecaptcha.ready(() => {
+                            resolve((window as any).grecaptcha)
+                        })
+                    } else {
+                        setTimeout(checkReady, 100)
+                    }
+                }
+                checkReady()
+            }
+
+            script.onerror = (): void => reject(new Error('Failed to load reCAPTCHA script.'))
+
+            document.head.appendChild(script)
+        })
+    }
+
+    const initCaptcha = () => {
+        loadRecaptchaScript()
+            .then((grecaptcha: any) => {
+                console.log('reCAPTCHA script loaded successfully.')
+
+                const recaptchaRes: RecaptchaContext = {
+                    render: (elementId: string, callback: (response: string) => void): void => {
+                        grecaptcha.ready(() => {
+                            try {
+                                grecaptcha.render(elementId, {
+                                    sitekey: siteKey,
+                                    callback,
+                                })
+                            } catch (error) {
+                                console.error('Error rendering reCAPTCHA:', error)
+                            }
+                        })
+                    },
+                    execute: (action: string, callback: (token: string) => void): void => {
+                        grecaptcha.ready(() => {
+                            try {
+                                grecaptcha.execute(siteKey, { action }).then((token: string) => {
+                                    callback(token)
+                                }).catch((error: any) => {
+                                    console.error('Error executing reCAPTCHA:', error)
+                                    throw error
+                                })
+                            } catch (error) {
+                                console.error('Error in reCAPTCHA execute:', error)
+                                throw error
+                            }
+                        })
+                    },
+                }
+                recaptcha.value = recaptchaRes
+            })
+            .catch((error: Error) => {
+                console.error('Error loading reCAPTCHA:', error)
+            })
+    }
 
     const getToken = async (action: string = 'submit'): Promise<string> => {
         return new Promise((resolve, reject) => {
@@ -11,13 +105,13 @@ export const useRecaptcha = () => {
                 return
             }
 
-            if (!$recaptcha) {
+            if (!recaptcha.value) {
                 reject(new Error('reCAPTCHA not loaded'))
                 return
             }
 
             try {
-                ($recaptcha as RecaptchaContext).execute(action, (token: string) => {
+                (recaptcha.value as RecaptchaContext).execute(action, (token: string) => {
                     resolve(token)
                 })
             } catch (error) {
@@ -27,11 +121,12 @@ export const useRecaptcha = () => {
     }
 
     const isLoaded = (): boolean => {
-        return (process.client && !!$recaptcha) as boolean
+        return (process.client && !!recaptcha.value) as boolean
     }
 
     return {
         getToken,
-        isLoaded
+        isLoaded,
+        initCaptcha
     }
 }
